Move tests into their own directory.
Create sub modules for cargo test to find the new tests. Add a simpler one-node suffix tree test. Add default binary.
