+++ Cargo.toml
@@ -2,6 +2,7 @@
 name = "rusty-plants"
 version = "0.1.0"
 edition = "2021"
+default-run = "rusty-plants"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
