+++ tests/tests.rs
@@ -1,0 +1,5 @@
+// Data structures tests
+mod data_structures {
+    mod edit_distance_tests;
+    mod suffix_tree_tests;
+}
